rename model to regressor and network to model
