+++ config/main.yaml
@@ -4,8 +4,8 @@ defaults:
   - proxy: dropout_regressor
   - oracle: grid_corners
   - dataset: dataset
-  - model: dropout_regressor
-  - network: mlp
+  - regressor: dropout_regressor
+  - model: mlp
 
 n_samples: 100
 al_n_rounds: 10
