+++ config/proxy/dropout_regressor.yaml
@@ -1,3 +1,3 @@
 defaults:
   - base
-_target_: proxy.acquisition.DropoutRegressor
+_target_: proxy.dropout_regressor.DropoutRegressor
